requirements.txt
streamlit>=1.28.0
pandas>=2.2.0
plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.13.0
flask>=2.3.3
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
